// src/webparts/simpleAnalytics/models/IAnalyticsStat.ts
export interface IAnalyticsStat {
    title: string;
    viewsLifetime: number;
    viewsRecent: number;
    viewsLast1Days: number;
    viewsLast2Days: number;
    viewsLast3Days: number;
    viewsLast4Days: number;
    viewsLast5Days: number;
    viewsLast6Days: number;
    viewsLast7Days: number;
    viewsLastMonths1: number;
    viewsLastMonths2: number;
    viewsLastMonths3: number;
    path: string;
    contentclass: string;
  }
  